changed the flag var to int type because byte is the same size actual memory.

+++ Part2/AsteroidsGame.pde
@@ -5,7 +5,7 @@ public static final int DOWN_BIT  = 2;
 public static final int RIGHT_BIT = 3;
 public static final int BRAKE_BIT = 4;
 public static final int HYPER_BIT = 5;
-public byte key_bits = 0x00;
+public int key_bits = 0x00;
 
 /**basic spaceship stuff**/
 public static final int basic_corners = 3;
@@ -55,22 +55,22 @@ public void draw()
   spacey.show();
 }
 public void scan_key_bits(){
-  if((key_bits & (byte)(1<<BRAKE_BIT))!=0){
+  if((key_bits & (1<<BRAKE_BIT))!=0){
     spacey.brake(0.1);
   }
-  if((key_bits & (byte)(1<<HYPER_BIT))!=0){
+  if((key_bits & (1<<HYPER_BIT))!=0){
     spacey.hyperspace();
   }
-  if((key_bits & (byte)(1<<UP_BIT))!=0){
+  if((key_bits & (1<<UP_BIT))!=0){
     spacey.accelerate((double)0.1);
   }
-  if((key_bits & (byte)(1<<DOWN_BIT))!=0){
+  if((key_bits & (1<<DOWN_BIT))!=0){
     spacey.accelerate((double)(-0.1));
   }
-  if((key_bits & (byte)(1<<LEFT_BIT))!=0){
+  if((key_bits & (1<<LEFT_BIT))!=0){
     spacey.rotate(-8);
   }
-  if((key_bits & (byte)(1<<RIGHT_BIT))!=0){
+  if((key_bits & (1<<RIGHT_BIT))!=0){
     spacey.rotate(8);
   }
 }
